moved into new file and can also search tv table
