+++ server/index.js
@@ -5,10 +5,10 @@ var pg = require('pg');
 var routes = express.Router();
 var sass = require('node-sass-endpoint');
 var User = require('./users.js');
-// var pgClient = require('./db.js');
-var sessions = require('./sessions.js')
-var cookieParser = require('cookie-parser')
-var combo = require('./combos.js')
+var sessions = require('./sessions.js');
+var cookieParser = require('cookie-parser');
+var combo = require('./combos.js');
+var suggestions = require('./suggestions.js');
 
 require('../db/seed/seedRestaurant.js');
 require('../db/seed/seedMovie.js');
@@ -39,10 +39,11 @@ routes.get('/css/app-bundle.css', sass.serve('./client/scss/app.scss'));
 
 routes.post('/signin', function(req, res, next){
   User.signin(req, res, next)
-})
+});
+
 routes.post('/signup', function(req, res, next){
   User.signup(req, res, next)
-})
+});
 
 routes.post('/saveWork', function(req, res, next){
   if (req.session){
@@ -50,58 +51,40 @@ routes.post('/saveWork', function(req, res, next){
   } else {
     res.status(401).send({Error: "User is not logged in"})
   }
-})
+});
+
 routes.post('/saveRestaurant', function(req, res, next){
   if (req.session){
     combo.saveRestaurant(req, res)
   } else {
     res.status(401).send({Error: "User is not logged in"})
   }  
-})
+});
+
 routes.post('/saveCombo', function(req, res, next){
   if (req.session){
     combo.saveCombo(req, res)
   } else {
     res.status(401).send({Error: "User is not logged in"})
   }
-})
+});
+
 routes.get('/userCombos', function(req, res, next){
   if (req.session){
     combo.pullCombos(req, res)
   } else {
     res.status(401).send({Error: "User is not logged in"})
   }
-})
+});
 //
 // Match endpoint to match movie genres with cuisines
 //
 routes.get('/api/match/:zip', function(req, res) {
-  var zip = req.params.zip;
-  // Get first 3 zip digits for SQL "like" query.
-  var slimZip = zip.slice(0,3);
-
-  var combinedResult = {};
-  var pgClient = new pg.Client(pgConConfig);
-  var restaurantQuery = pgClient.query("SELECT * FROM restaurants WHERE restaurant_zip LIKE '" + slimZip + "%' order by random() limit 1", function(err, result){
-    return result;
-  });
-  restaurantQuery.on('end', function(result) {
-    combinedResult.restaurant = result.rows[0];
-  });
-
-  var movieQuery = pgClient.query("SELECT * FROM movies order by random() limit 1", function(err, result){
-    return result;
-  });
-  movieQuery.on('end', function(result) {
-    combinedResult.movie = result.rows[0];
-    res.send(combinedResult)
-  });
-  pgClient.on('drain', function() {
-  pgClient.end();
-});
-
-pgClient.connect()
-
+  if (req.body.type === 'TV') {
+    suggestions.getTVSuggestions(req, res)
+  } else {
+    suggestions.getMovieSuggestion(req, res)
+  }
 });
 
 //
